+++ zenmaster/errors.py
@@ -1,0 +1,14 @@
+class ZenMasterError(RuntimeError):
+    pass
+
+
+class BackendUnavailable(ZenMasterError):
+    pass
+
+
+class SMUNotInitialized(ZenMasterError):
+    pass
+
+
+class UnsupportedCPU(ZenMasterError):
+    pass